Renamed inih files to avoid conflicts and added it to the include path
